分离请求类
