+++ src/Comm/Pool.php
@@ -14,6 +14,7 @@ class Pool
 
     /**
      * 公共缓存池
+     * 以文件为单位区分key
      * 结构如下：[类名 => [$key => $value]]
      * [
      *     'Db' => [
